src: Correctness fixes, cleanups & GPU blocklist
Sadly PowerVR as on the JingPad has serious trouble dumping EGLImages
into an OpenGL texture when the image reaches a certain size, most
noticable in Lomiri with the Wallpaper component being corrupted.
Chromium bugs indicate similar issues with PowerVR and EGLImage, so
disallow usage of GrallocTextures on it and fall back to the default.